requirements.txt
pandas
numpy
mysql-connector-python
python-dotenv
xgboost
scikit-learn
matplotlib
